FC: the mini skips the piezo it does not have, so turning sounds on no longer aborts it at every boot (#1481)
board_m1.h declares PIEZO_PIN = -1, but setup_fc configured it and every
piezo write went to it anyway. The IDF GPIO driver rejects -1 with an
ESP_LOGE per call: two error lines at every boot.

With sounds on it is worse. A beep's first write comes from piezoStart()
inside portENTER_CRITICAL (the #732 serialisation). On the S3 that masks
interrupts to the EXCM level. The error log's vprintf then takes
picolibc's recursive stdout lock, and lock_acquire_generic() calls
abort() when it cannot yield. The boot chirp fires about a second after
every boot, and SOUNDS_ENABLE stores the flag before its confirmation
beep, so turning sounds on would reboot the mini's flight computer about
a second into every boot. Traced through the IDF 6.0 source and the M1
image; not observed on hardware.

Now a negative PIEZO_PIN means no piezo. setup_fc skips the GPIO and the
timer (the SERVO_ACT_PIN / GPS_ACT_PIN idiom), so piezo_pwm_ready stays
false and every beep returns early in piezoStart(). piezoApply() returns
on a negative pin. Every write after setup goes through it, piezoStop()
included. It keys on the pin, not on piezo_pwm_ready, so a board with a
coil whose timer failed still gets the unconditional drive-low.

PIEZO_PIN is an int on V7/V8/V9 too. GCC 15 warns on a uint8_t >= 0
(-Wtype-limits, even through a cast), and a uint8_t cannot hold the -1
that V10 will need once its piezo moves to the out computer (#1409).

V9 is unchanged: the piezo functions and setup_fc compile to identical
instructions. On M1 the piezo's three GPIO call sites are gone and
piezoStart() is empty.

Refs #1409

+++ tinkerrocket-idf/projects/flight_computer/main/board/board_v7.h
@@ -85,7 +85,7 @@ struct board_pins
     static constexpr uint8_t SERVO_PIN_4 = 46;
 
     // ### Indicators ###
-    static constexpr uint8_t PIEZO_PIN = 53;
+    static constexpr int PIEZO_PIN = 53;
     static constexpr uint8_t RED_LED_PIN = 2;
     static constexpr uint8_t BLUE_LED_PIN = 29;
